+++ CHANGELOG.md
@@ -2,6 +2,7 @@ v1.9.0
 
 - Much faster rendering
 - Add support for http-tiles data in track config
+- Faster track resizing via y axis scaling
 
 v1.8.1